add inference for liu and wang

+++ repath/preprocess/patching/patch_index.py
@@ -131,7 +131,7 @@ class CombinedIndex(object):
         ci = cls(cps)
         return ci
 
-    def save_patches(self, output_dir: Path, transforms: List[transforms.Compose] = None, affix: str = '') -> None:
+    def save_patches(self, output_dir: Path, transforms: List[transforms.Compose] = None, affix: str = '', add_patches: bool = False) -> None:
         """Saves patches after applying transforms to them for more than one datastes.
 
         Args:
@@ -139,18 +139,19 @@ class CombinedIndex(object):
             transforms (List[transforms.Compose], optional): List of transforms to be applied to each patch. Defaults to None.
             affix (str, optional): A string added to the name of the patch before saving it. Defaults to ''.
         """
-        # check subdirectories exist and delete any existing patches
-        for cps_idx, cps_group in self.patches_df.groupby('cps_idx'):
-            # get the patch label as a string
-            labels_dict = remove_item_from_dict(self.datasets[cps_idx].labels, 'background')
-            labels = list(labels_dict.keys())
-            for lab in labels:
-                # ensure the output directory exists
-                output_subdir = output_dir / lab
-                output_subdir.mkdir(parents=True, exist_ok=True)
-                files_in_subdir = output_subdir.glob('*.png')
-                for ff in files_in_subdir:
-                    os.remove(ff)
+        if not add_patches:
+            # check subdirectories exist and delete any existing patches
+            for cps_idx, cps_group in self.patches_df.groupby('cps_idx'):
+                # get the patch label as a string
+                labels_dict = remove_item_from_dict(self.datasets[cps_idx].labels, 'background')
+                labels = list(labels_dict.keys())
+                for lab in labels:
+                    # ensure the output directory exists
+                    output_subdir = output_dir / lab
+                    output_subdir.mkdir(parents=True, exist_ok=True)
+                    files_in_subdir = output_subdir.glob('*.png')
+                    for ff in files_in_subdir:
+                        os.remove(ff)
 
         for cps_idx, cps_group in self.patches_df.groupby('cps_idx'):
             for slide_idx, sl_group in cps_group.groupby('slide_idx'):
